monitor_vykonu/requirements.txt
snap7
mysql-connector-python
smtplib
struct
socket
logging
xml.etree.ElementTree
